backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9
groq>=0.15.0
pymupdf==1.24.0
Pillow==10.2.0
pydantic==2.6.4
pydantic-settings==2.2.1
python-dotenv==1.0.1
reportlab==4.1.0
httpx<=0.27.2
